Smart Code Tutor/backend/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
python-dotenv==1.1.1
e2b-code-interpreter==1.5.2
openai==1.95.0
python-socketio==5.13.0
python-engineio==4.12.2
pydantic==2.11.7
numpy==1.26.4 